+++ doc/source/changes.rst
@@ -7,6 +7,9 @@ Changes
 New in 0.16
 -----------
 
+* Epoch definitions in :attr:`MneExperiment.epochs` can now include a ``'base'``
+  parameter, which will copy the given "base" epoch and modify it with the
+  current definition.
 * :meth:`MneExperiment.make_mov_ttest` and
   :meth:`MneExperiment.make_mov_ga_dspm` are fixed but require PySurfer 0.6.
 * API: :mod:`plot.brain` function signatures changed slightly to accommodate